rename config to const
