+++ evesso/authorize.py
@@ -6,7 +6,7 @@ from urllib.parse import urlencode, urlparse, parse_qs
 
 import requests
 
-from .config import AUTH_URL, TOKEN_URL, HEADERS
+from .const import AUTH_URL, TOKEN_URL, HEADERS
 from .callback import listen_for_callback
 
 STATE = secrets.token_urlsafe(32)
